Order the nodes color domains
Based off the distance to a source node with that color. Closer
source nodes will be first in the domain, and will be tried first
in the apple algorithm.

This can now solve the 14x14 board.

I am going to call apple the dumb one and banana the smart one.